requirements.txt
pg8000==1.11.0
python-telegram-bot==12.0.0b1
Flask-SQLAlchemy==2.3.1
gunicorn==19.9.0
